cleaned up fixture support
