+++ java/src/test/scala/com/logicalclocks/hsfs/utils/FixtureSupport.scala
@@ -16,7 +16,7 @@
 
 package com.logicalclocks.hsfs.utils
 
-import org.apache.spark.sql.types.{DoubleType, LongType, MapType, StringType, StructType}
+import org.apache.spark.sql.types.{ArrayType, BinaryType, BooleanType, ByteType, DateType, Decimal, DecimalType, DoubleType, FloatType, IntegerType, LongType, MapType, ShortType, StringType, StructField, StructType, TimestampType}
 import org.apache.spark.sql.{DataFrame, Row, SparkSession}
 
 import scala.util.Random
@@ -24,353 +24,160 @@ import scala.util.Random
 
 trait FixtureSupport {
 
-  def getEmptyColumnDataDf(sparkSession: SparkSession): DataFrame = {
+  def getDfWithDifferentDatatypes(sparkSession: SparkSession): DataFrame = {
     import sparkSession.implicits._
 
-    Seq(
-      ("", "a", "f"),
-      ("", "b", "d"),
-      ("", "a", null),
-      ("", "a", "f"),
-      ("", "b", null),
-      ("", "a", "f")
-    ).toDF("att1", "att2", "att3")
-  }
-
-  def getDfEmpty(sparkSession: SparkSession): DataFrame = {
-    import sparkSession.implicits._
-    val column1 = $"column1".string
-    val column2 = $"column2".string
-    val mySchema = StructType(column1 :: column2 :: Nil)
-
-    sparkSession.createDataFrame(sparkSession.sparkContext.emptyRDD[Row], mySchema)
-  }
-
-  def getDfWithNRows(sparkSession: SparkSession, n: Int): DataFrame = {
-    import sparkSession.implicits._
-
-    (1 to n)
-      .toList
-      .map { index => (s"$index", s"c1-r$index", s"c2-r$index")}
-      .toDF("c0", "c1", "c2")
-  }
-
-  def getDfWithNestedColumn(sparkSession: SparkSession): DataFrame = {
-
-    import sparkSession.implicits._
-
-//    val schema = new StructType()
-//      .add("dc_id", StringType)
-//      .add("lang", StringType)
-//      .add("source",
-//        MapType(
-//          StringType,
-//          new StructType()
-//            .add("description", StringType)
-//            .add("weblink", StringType)
-//            .add("ip", StringType)
-//            .add("id", LongType)
-//            .add("temp", LongType)
-//            .add("c02_level", LongType)
-//            .add("geo",
-//              new StructType()
-//                .add("lat", DoubleType)
-//                .add("long", DoubleType)
-//            )
-//        )
-//      )
-
-    // Create a single entry with id and its complex and nested data types
-    Seq("""
-      {
-      "dc_id": "dc-101",
-      "lang":"en-us",
-      "source": {
-          "sensor-igauge": {
-            "id": 10,
-            "ip": "68.28.91.22",
-            "description": "Sensor attached to the container ceilings",
-            "weblink": "http://www.testing-url.dk/",
-            "temp":35,
-            "c02_level": 1475,
-            "geo": {"lat":38.00, "long":97.00}
-            }
-          }
-        }
-      }""").toDF()
-  }
-
-
-  def getDfMissing(sparkSession: SparkSession): DataFrame = {
-    import sparkSession.implicits._
-
-    Seq(
-      ("1", "a", "f"),
-      ("2", "b", "d"),
-      ("3", null, "f"),
-      ("4", "a", null),
-      ("5", "a", "f"),
-      ("6", null, "d"),
-      ("7", null, "d"),
-      ("8", "b", null),
-      ("9", "a", "f"),
-      ("10", null, null),
-      ("11", null, "f"),
-      ("12", null, "d")
-    ).toDF("item", "att1", "att2")
-  }
-
-  def getDfFull(sparkSession: SparkSession): DataFrame = {
-    import sparkSession.implicits._
-
-    Seq(
-      ("1", "a", "c"),
-      ("2", "a", "c"),
-      ("3", "a", "c"),
-      ("4", "b", "d")
-    ).toDF("item", "att1", "att2")
-  }
-
-  def getDfWithNegativeNumbers(sparkSession: SparkSession): DataFrame = {
-    import sparkSession.implicits._
-
-    Seq(
-      ("1", "-1", "-1.0"),
-      ("2", "-2", "-2.0"),
-      ("3", "-3", "-3.0"),
-      ("4", "-4", "-4.0")
-    ).toDF("item", "att1", "att2")
-  }
-
-  def getDfCompleteAndInCompleteColumns(sparkSession: SparkSession): DataFrame = {
-    import sparkSession.implicits._
-
-    Seq(
-      ("1", "a", "f"),
-      ("2", "b", "d"),
-      ("3", "a", null),
-      ("4", "a", "f"),
-      ("5", "b", null),
-      ("6", "a", "f")
-    ).toDF("item", "att1", "att2")
-  }
-
-  def getDfCompleteAndInCompleteColumnsDelta(sparkSession: SparkSession): DataFrame = {
-    import sparkSession.implicits._
-
-    Seq(
-      ("7", "a", null),
-      ("8", "b", "d"),
-      ("9", "a", null)
-    ).toDF("item", "att1", "att2")
-  }
-
-
-  def getDfFractionalIntegralTypes(sparkSession: SparkSession): DataFrame = {
-    import sparkSession.implicits._
-
-    Seq(
-      ("1", "1.0"),
-      ("2", "1")
-    ).toDF("item", "att1")
-  }
-
-  def getDfFractionalStringTypes(sparkSession: SparkSession): DataFrame = {
-    import sparkSession.implicits._
-
-    Seq(
-      ("1", "1.0"),
-      ("2", "a")
-    ).toDF("item", "att1")
-  }
-
-  def getDfIntegralStringTypes(sparkSession: SparkSession): DataFrame = {
-    import sparkSession.implicits._
-
-    Seq(
-      ("1", "1"),
-      ("2", "a")
-    ).toDF("item", "att1")
-  }
-
-  def getDfWithNumericValues(sparkSession: SparkSession): DataFrame = {
-    import sparkSession.implicits._
-    // att2 is always bigger than att1
-    Seq(
-      ("1", 1, 0, 0),
-      ("2", 2, 0, 0),
-      ("3", 3, 0, 0),
-      ("4", 4, 5, 4),
-      ("5", 5, 6, 6),
-      ("6", 6, 7, 7)
-    ).toDF("item", "att1", "att2", "att3")
-  }
-
-  def getDfWithNumericFractionalValues(sparkSession: SparkSession): DataFrame = {
-    import sparkSession.implicits._
-    Seq(
-      ("1", 1.0, 0.0),
-      ("2", 2.0, 0.0),
-      ("3", 3.0, 0.0),
-      ("4", 4.0, 5.0),
-      ("5", 5.0, 6.0),
-      ("6", 6.0, 7.0)
-    ).toDF("item", "att1", "att2")
-  }
-
-  def getDfWithDecimalFractionalValues(sparkSession: SparkSession): DataFrame = {
-    import sparkSession.implicits._
-    Seq(
-      ("1", BigDecimal(1.0), BigDecimal(0.0)),
-      ("2", BigDecimal(2.0), BigDecimal(0.0)),
-      ("3", BigDecimal(3.0), BigDecimal(0.0)),
-      ("4", BigDecimal(4.0), BigDecimal(5.0)),
-      ("5", BigDecimal(5.0), BigDecimal(6.0)),
-      ("6", BigDecimal(6.0), BigDecimal(7.0))
-    ).toDF("item", "att1", "att2")
-  }
-
-  def getDfWithNumericFractionalValuesForKLL(sparkSession: SparkSession): DataFrame = {
-    import sparkSession.implicits._
-    Seq(
-      ("1", 1.0, 0.0),
-      ("2", 2.0, 0.0),
-      ("3", 3.0, 0.0),
-      ("4", 4.0, 5.0),
-      ("5", 5.0, 6.0),
-      ("6", 6.0, 7.0),
-      ("7", 7.0, 0.0),
-      ("8", 8.0, 0.0),
-      ("9", 9.0, 0.0),
-      ("10", 10.0, 5.0),
-      ("11", 11.0, 6.0),
-      ("12", 12.0, 7.0),
-      ("13", 13.0, 0.0),
-      ("14", 14.0, 0.0),
-      ("15", 15.0, 0.0),
-      ("16", 16.0, 5.0),
-      ("17", 17.0, 6.0),
-      ("18", 18.0, 7.0),
-      ("19", 19.0, 0.0),
-      ("20", 20.0, 0.0),
-      ("21", 21.0, 0.0),
-      ("22", 22.0, 5.0),
-      ("23", 23.0, 6.0),
-      ("24", 24.0, 7.0),
-      ("25", 25.0, 0.0),
-      ("26", 26.0, 0.0),
-      ("27", 27.0, 0.0),
-      ("28", 28.0, 5.0),
-      ("29", 29.0, 6.0),
-      ("30", 30.0, 7.0)
-    ).toDF("item", "att1", "att2")
-  }
-
-  def getDfWithNumericFractionalValuesForPMF(sparkSession: SparkSession): DataFrame = {
-    import sparkSession.implicits._
-    Seq(
-      ("1", 1.0, 6.0),
-      ("2", 2.0, 6.0),
-      ("3", 3.0, 6.0),
-      ("4", 4.0, 6.0),
-      ("5", 5.0, 6.0),
-      ("6", 6.0, 7.0),
-      ("7", 7.0, 0.0),
-      ("8", 8.0, 0.0),
-      ("9", 9.0, 0.0),
-      ("10", 10.0, 5.0),
-      ("11", 11.0, 6.0),
-      ("12", 12.0, 7.0),
-      ("13", 13.0, 0.0),
-      ("14", 14.0, 0.0),
-      ("15", 15.0, 0.0),
-      ("16", 16.0, 5.0),
-      ("17", 17.0, 6.0),
-      ("18", 18.0, 7.0),
-      ("19", 19.0, 0.0),
-      ("20", 20.0, 0.0),
-      ("21", 21.0, 0.0),
-      ("22", 22.0, 5.0),
-      ("23", 23.0, 6.0),
-      ("24", 24.0, 7.0),
-      ("25", 25.0, 0.0),
-      ("26", 26.0, 0.0),
-      ("27", 27.0, 0.0),
-      ("28", 28.0, 5.0),
-      ("29", 29.0, 6.0),
-      ("30", 30.0, 7.0)
-    ).toDF("item", "att1", "target")
-  }
-
-  def getDfWithUniqueColumns(sparkSession: SparkSession): DataFrame = {
-    import sparkSession.implicits._
-
-    Seq(
-      ("1", "0", "3", "1", "5", "0"),
-      ("2", "0", "3", "2", "6", "0"),
-      ("3", "0", "3", null, "7", "0"),
-      ("4", "5", null, "3", "0", "4"),
-      ("5", "6", null, "4", "0", "5"),
-      ("6", "7", null, "5", "0", "6")
+    val schema = StructType(
+      StructField("bool", BooleanType, true) ::
+        StructField("byte", ByteType, true) ::
+        StructField("short", ShortType, true) ::
+        StructField("int", IntegerType, true) ::
+        StructField("long", LongType, true) ::
+        StructField("float", FloatType, true) ::
+        StructField("dbl", DoubleType, true) ::
+        StructField("dec", DecimalType(38, 28), true) ::
+        StructField("ts", TimestampType, true) ::
+        StructField("dt", DateType, true) ::
+        StructField("str", StringType, true) ::
+        StructField("bn", BinaryType, true) ::
+        StructField("arr", ArrayType(IntegerType, true), true) ::
+        StructField("map", MapType(StringType, StringType, true), true) ::
+        StructField("struct", StructType(
+          List(
+            StructField("favorite_color", StringType, true),
+            StructField("age", IntegerType, true)
+          )
+        ), true) :: Nil
     )
-    .toDF("unique", "nonUnique", "nonUniqueWithNulls", "uniqueWithNulls",
-      "onlyUniqueWithOtherNonUnique", "halfUniqueCombinedWithNonUnique")
+
+    import java.io.{ByteArrayOutputStream, ObjectOutputStream}
+    val serialise = (value: Any) => {
+      val stream: ByteArrayOutputStream = new ByteArrayOutputStream()
+      val oos = new ObjectOutputStream(stream)
+      oos.writeObject(value)
+      oos.close()
+      stream.toByteArray
+    }
+
+    val dataList = Seq(
+      Row(
+        true,
+        7.toByte,
+        15.toShort,
+        3743,
+        327828732L,
+        5F,
+        123.5,
+        Decimal("1208484888.8474763788847476378884747637"),
+        java.sql.Timestamp.valueOf("2020-06-29 22:41:30"),
+        java.sql.Date.valueOf("2020-06-29"),
+        "str",
+        serialise("strBinary"),
+        Array(17, 2, 3),
+        Map("aguila" -> "Colombia", "modelo" -> "Mexico"),
+        Row("blue", 45)),
+      Row(
+        false,
+        8.toByte,
+        9.toShort,
+        3742,
+        32728732L,
+        10F,
+        12.5,
+        Decimal("1208484889.8474763788847476378884747637"),
+        java.sql.Timestamp.valueOf("2020-06-30 22:41:30"),
+        java.sql.Date.valueOf("2020-06-30"),
+        "str2",
+        serialise("strBinary2"),
+        Array(17, 2, 3),
+        Map("aguila" -> "Colombia2", "modelo" -> "Mexico2"),
+        Row("brown", 46)),
+      Row(
+        true,
+        7.toByte,
+        15.toShort,
+        3743,
+        327828732L,
+        5F,
+        123.5,
+        Decimal("1208484888.8474763788847476378884747637"),
+        java.sql.Timestamp.valueOf("2020-06-29 22:41:30"),
+        java.sql.Date.valueOf("2020-06-29"),
+        "str",
+        serialise("strBinary"),
+        Array(1, 2, 7),
+        Map("aguila" -> "Colombia", "modelo" -> "Mexico"),
+        Row("blue", 45)),
+      Row(
+        false,
+        8.toByte,
+        9.toShort,
+        3742,
+        32728732L,
+        0F,
+        12.5,
+        Decimal("1208484889.8474763788847476378884747637"),
+        java.sql.Timestamp.valueOf("2020-06-30 22:41:30"),
+        java.sql.Date.valueOf("2020-06-30"),
+        "str2",
+        serialise("strBinary2"),
+        Array(1, 2, 7),
+        Map("aguila" -> "Colombia2", "modelo" -> "Mexico2"),
+        Row("brown", 46)),
+      Row(
+        null,
+        null,
+        null,
+        null,
+        null,
+        null,
+        null,
+        null,
+        null,
+        null,
+        null,
+        null,
+        null,
+        null,
+        null),
+      Row(
+        false,
+        2.toByte,
+        22.toShort,
+        37342,
+        327228732L,
+        Float.NaN,
+        Double.NegativeInfinity,
+        Decimal("1228384889.8474763788847476378884747637"),
+        java.sql.Timestamp.valueOf("2020-07-30 22:41:30"),
+        java.sql.Date.valueOf("2020-07-30"),
+        "str3",
+        serialise("strBinary3"),
+        Array(1, 2),
+        Map("aguila" -> "Colombia2", "modelo" -> "Mexico3"),
+        Row("brown", 47)),
+      Row(
+        false,
+        2.toByte,
+        22.toShort,
+        37342,
+        327228732L,
+        Float.PositiveInfinity,
+        Double.NegativeInfinity,
+        Decimal("1228384889.8474763788847476378884747637"),
+        java.sql.Timestamp.valueOf("2020-07-30 22:41:30"),
+        java.sql.Date.valueOf("2020-07-30"),
+        "str3",
+        serialise("strBinary3"),
+        Array(1, 2),
+        Map("aguila" -> "Colombia2", "modelo" -> "Mexico3"),
+        Row("brown", 47))
+    )
+
+    sparkSession.createDataFrame(
+      sparkSession.sparkContext.parallelize(dataList),
+      schema
+    )
   }
 
-  def getDfWithDistinctValues(sparkSession: SparkSession): DataFrame = {
-    import sparkSession.implicits._
-
-    Seq(
-      ("a", null),
-      ("a", null),
-      (null, "x"),
-      ("b", "x"),
-      ("b", "x"),
-      ("c", "y"))
-      .toDF("att1", "att2")
-  }
-
-  def getDfWithConditionallyUninformativeColumns(sparkSession: SparkSession): DataFrame = {
-    import sparkSession.implicits._
-    Seq(
-      (1, 0),
-      (2, 0),
-      (3, 0)
-    ).toDF("att1", "att2")
-  }
-
-  def getDfWithConditionallyInformativeColumns(sparkSession: SparkSession): DataFrame = {
-    import sparkSession.implicits._
-    Seq(
-      (1, 4),
-      (2, 5),
-      (3, 6)
-    ).toDF("att1", "att2")
-  }
-
-  def getDfWithCategoricalColumn(
-      sparkSession: SparkSession,
-      numberOfRows: Int,
-      categories: Seq[String])
-    : DataFrame = {
-
-    val random = new Random(0)
-
-    import sparkSession.implicits._
-    (1 to numberOfRows)
-      .toList
-      .map { index => (s"$index", random.shuffle(categories).head)}
-      .toDF("att1", "categoricalColumn")
-  }
-
-  def getDfWithVariableStringLengthValues(sparkSession: SparkSession): DataFrame = {
-    import sparkSession.implicits._
-    Seq(
-      "",
-      "a",
-      "bb",
-      "ccc",
-      "dddd"
-    ).toDF("att1")
-  }
 }
